requirements.txt
fastapi
sqlalchemy
uvicorn
psycopg2
python-multipart
bcrypt
dotenv
pyjwt